requirements.txt
Django==5.0
djangorestframework==3.15.2
celery==5.3.6
redis==5.0.4
psycopg2-binary==2.9.9
python-decouple==3.8
httpx==0.27.0
openai==1.30.0
semgrep==1.75.0
bandit==1.7.9
trufflehog3==3.0.9
trivy==0.51.0
sslyze==5.2.0
corsy==0.1
wapiti3==3.1.8
sqlmap==1.8.4
coverage==7.4.4
python-dotenv==1.0.1
django-environ==0.11.2
dj-database-url==2.1.0

